+++ src/components/NewsStand/Content/NewsCompany/CompanyInfoCard/CompanyInfoCard.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useRecoilState } from 'recoil';
-import { myNewsSubscribeState } from '@recoil/news';
+import { myNewsSubscribeState } from '@recoilStore/news';
 import { Card, CardBtn } from './CompanyInfoCard.style';
 import { COMPANY_CARD } from '@utils/constant';
 
@@ -8,15 +8,9 @@ const CompanyInfoCard = ({ id, company, logoImgUrl }) => {
   const [myNewsStandList, setMyNewsStandList] = useRecoilState(myNewsSubscribeState);
   const isSubscribeCompany = myNewsStandList.some((subscribeId) => subscribeId === id);
 
-  const handleSubscribe = () => {
-    setMyNewsStandList((subscribeList) => [...subscribeList, id]);
-  };
+  const handleSubscribe = () => setMyNewsStandList((subscribeList) => [...subscribeList, i]);
 
-  const handleUnSubscribe = () => {
-    setMyNewsStandList((list) => {
-      return list.filter((subscribeId) => subscribeId !== id);
-    });
-  };
+  const handleUnSubscribe = () => setMyNewsStandList((list) => list.filter((subscribeId) => subscribeId !== id));
 
   const SubScribeBtn = () => {
     if (isSubscribeCompany) {
