--- src/util/Router/Route.jsx
@@ -1,10 +1,0 @@
-import { useContext } from 'react';
-import { HistoryContext } from './Router';
-
-export const Route = ({ path, children, ...props }) => {
-  const { currentPath } = useContext(HistoryContext);
-  if (path !== currentPath) {
-    return null;
-  }
-  return <div {...props}>{children}</div>;
-};
