--- src/util/Router/index.js
@@ -1,5 +1,0 @@
-export * from './Link';
-export * from './NotFound';
-export * from './Route';
-export * from './routes';
-export * from './Router';
