--- src/util/Router/routes.js
@@ -1,11 +1,0 @@
-export const routes = {
-  home: {
-    path: '/',
-  },
-  labels: {
-    path: '/labels',
-  },
-  milestones: {
-    path: '/milestones',
-  },
-};
